Refactored to pull error log into its own class.
