+++ src/main/java/com/hmsonline/cassandra/triggers/TestTrigger.java
@@ -10,6 +10,7 @@ public class TestTrigger implements Trigger {
     public void process(LogEntry logEntry) {
         logger.debug("Trigger processing : [" + logEntry.getUuid() + "]");   
         wasCalled = true;
+        throw new RuntimeException("Foo");
     }
 
     public static boolean wasCalled() {
